Split route tests between prepareRoutes & createRouter
